Move `INotificationQueue` from `atyping` to the imlementation file.
